==> analysis/03-MarkerAnalysisBCT.Rmd <==
---
title: "Marker analysis of broad cell types"
author: "Belinda Phipson & Neda R. Mehdiabadi"
date: "10/02/2021-27/09/2021"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
    echo = TRUE,
    message = FALSE,
    warning = FALSE
)
```

# Load libraries and functions

```{r}
library(edgeR)
library(RColorBrewer)
library(org.Hs.eg.db)
library(limma)
library(Seurat)
library(cowplot)
library(DelayedArray)
library(scran)
library(NMF)
library(workflowr)
library(ggplot2)
library(clustree)
library(dplyr)
library(gridBase)
library(grid)
library(speckle)
```


# Read in the data objects

```{r readTargets}
targets <- read.delim("data/targets.txt",header=TRUE, stringsAsFactors = FALSE)
targets$FileName2 <- paste(targets$FileName,"/",sep="")
targets$Group_ID2 <- gsub("LV_","",targets$Group_ID)
group <- c("fetal_1","fetal_2","fetal_3",
           "non-diseased_1","non-diseased_2","non-diseased_3",
           "diseased_1","diseased_2",
           "diseased_3","diseased_4")
m <- match(group, targets$Group_ID2)
targets <- targets[m,]
```

```{r}
# Load unfiltered counts matrix for every sample (object all)
load("/group/card2/Neda/MCRI_LAB/single_cell_nuclei_rnaseq/Porello-heart-snRNAseq/output/RDataObjects/all-counts.Rdata")
```

```{r}
heart.integrated <- readRDS("/group/card2/Neda/MCRI_LAB/must-do-projects/EnzoPorrelloLab/dilated-cardiomyopathy/data/heart-int-FND-filtered.Rds")
```

```{r}
Idents(heart.integrated) <- heart.integrated$Broad_celltype
heart.integrated$Broad_celltype <- factor(heart.integrated$Broad_celltype, levels = c("Er","CM(Prlf)","CM","Endo","Pericyte","Fib","Immune","Neuron","Smc"))

heart.integrated$biorep <- factor(heart.integrated$biorep,levels=c("f1","f2","f3","nd1","nd2","nd3","d1","d2","d3","d4"))
table(heart.integrated$biorep,heart.integrated$Broad_celltype)

all.bct <- factor(heart.integrated$Broad_celltype,
              levels = c("Er","CM(Prlf)","CM","Endo","Pericyte","Fib","Immune","Neuron","Smc"))
sample <- heart.integrated$biorep
```

# Get gene annotation and perform filtering

```{r}
columns(org.Hs.eg.db)
ann <- AnnotationDbi:::select(org.Hs.eg.db,keys=rownames(all),columns=c("SYMBOL","ENTREZID","ENSEMBL","GENENAME","CHR"),keytype = "SYMBOL")
m <- match(rownames(all),ann$SYMBOL)
ann <- ann[m,]
```

```{r}
mito <- grep("mitochondrial",ann$GENENAME)
ribo <- grep("ribosomal",ann$GENENAME)
missingEZID <- which(is.na(ann$ENTREZID))
```

```{r markers}
m <- match(colnames(heart.integrated),colnames(all))
all.counts <- all[,m]
```

```{r}
chuck <- unique(c(mito,ribo,missingEZID))
all.counts.keep <- all.counts[-chuck,]
ann.keep <- ann[-chuck,]
```

# remove x y genes
```{r}
xy <- ann.keep$CHR %in% c("X","Y")
all.counts.keep <- all.counts.keep[!xy,]
ann.keep <- ann.keep[!xy,]
```

```{r}
numzero.genes <- rowSums(all.counts.keep==0)
keep.genes <- numzero.genes < (ncol(all.counts.keep)-20)
all.keep <- all.counts.keep[keep.genes,]
ann.keep.all <- ann.keep[keep.genes,]
```


# Limma analysis

```{r}

logcounts.all <- normCounts(all.keep,log=TRUE,prior.count=0.5)

design <- model.matrix(~0+all.bct+sample)
colnames(design)[1:(length(levels(all.bct)))] <- levels(all.bct)

mycont <- matrix(0,ncol=length(levels(all.bct)),nrow=length(levels(all.bct)))
colnames(mycont)<-levels(all.bct)
diag(mycont)<-1
mycont[upper.tri(mycont)]<- -1/(length(levels(all.bct))-1)
mycont[lower.tri(mycont)]<- -1/(length(levels(all.bct))-1)

# Fill out remaining rows with 0s
zero.rows <- matrix(0,ncol=length(levels(all.bct)),nrow=(ncol(design)-length(levels(all.bct))))
test <- rbind(mycont,zero.rows)

fit <- lmFit(logcounts.all,design)
fit.cont <- contrasts.fit(fit,contrasts=test)
fit.cont <- eBayes(fit.cont,trend=TRUE,robust=TRUE)

fit.cont$genes <- ann.keep.all

treat.all <- treat(fit.cont,lfc=0.5)

```

```{r echo=FALSE}
#save(treat.all, file = "../output/bct-treat-all.RData")
#load("./output/bct-treat-all.RData")
```

```{r}
dt <- decideTests(treat.all)
summary(dt)
```


```{r,fig.width=10,fig.height=10}
par(mfrow=c(3,3))
par(mar=c(5,5,2,2))
for(i in 1:ncol(treat.all)){
  plotMD(treat.all,coef=i,status = dt[,i],hl.cex=0.5)
  abline(h=0,col=colours()[c(226)])
  lines(lowess(treat.all$Amean,treat.all$coefficients[,i]),lwd=1.5,col=4)
}
```


# DotPlot to visualise marker genes

```{r}
DefaultAssay(heart.integrated) <- "RNA"

sig.genes <- gene.label <- vector("list", ncol(treat.all))
for(i in 1:length(sig.genes)){
  top <- topTreat(treat.all,coef=i,n=Inf,sort.by="t")
  sig.genes[[i]] <- rownames(top)[top$logFC>0][1:10]
  gene.label[[i]] <- paste(rownames(top)[top$logFC>0][1:10],colnames(treat.all)[i],sep="-")
} 

csig <- unlist(sig.genes)
genes <- unlist(gene.label)

missing <- is.na(match(csig,rownames(heart.integrated)))

csig2 <- csig[!missing]

gene.cols <- rep(c(ggplotColors(8),"grey"),each=10)
gene.cols <- gene.cols[!missing]

d <- duplicated(csig2)
csig2 <- csig2[!d]
gene.cols <- gene.cols[!d]
```

```{r, fig.width=15, fig.height=15}
DotPlot(heart.integrated,features=unique(csig2),group.by="Broad_celltype",cols = c("lightgrey", "red"))+RotatedAxis() + FontSize(y.text = 8, x.text=14) + labs(y=element_blank(),x=element_blank()) + coord_flip() + theme(axis.text.y = element_text(color=(gene.cols)))
```

# Perform gene set testing on reactome sets

```{r}
load("/group/card2/Neda/MCRI_LAB/single_cell_nuclei_rnaseq/Porello-heart-snRNAseq/output/RDataObjects/human_c2_v5p2.rdata")
c2.id <- ids2indices(Hs.c2,treat.all$genes$ENTREZID)
reactome.id <-c2.id[grep("REACTOME",names(c2.id))]
```

## Reactome figure
Top to bottom: Er, CM(Prlf), CM, Endo, Pericyte, Fib, Immune, Neuron, Smc

```{r}
eryth.camera <- cameraPR(treat.all$t[,1],reactome.id)
eryth.camera.up <- eryth.camera[eryth.camera[,2]=="Up",]

cardioprlf.camera <- cameraPR(treat.all$t[,2],reactome.id)
cardioprlf.camera.up <- cardioprlf.camera[cardioprlf.camera[,2]=="Up",]

cardio.camera <- cameraPR(treat.all$t[,3],reactome.id)
cardio.camera.up <- cardio.camera[cardio.camera[,2]=="Up",]

endo.camera <- cameraPR(treat.all$t[,4],reactome.id)
endo.camera.up <- endo.camera[endo.camera[,2]=="Up",]

pericyte.camera <- cameraPR(treat.all$t[,5],reactome.id)
pericyte.camera.up <- pericyte.camera[pericyte.camera[,2]=="Up",]

fibro.camera <- cameraPR(treat.all$t[,6],reactome.id)
fibro.camera.up <- fibro.camera[fibro.camera[,2]=="Up",]

immune.camera <- cameraPR(treat.all$t[,7],reactome.id)
immune.camera.up <- immune.camera[immune.camera[,2]=="Up",]

neuron.camera <- cameraPR(treat.all$t[,8],reactome.id)
neuron.camera.up <- neuron.camera[neuron.camera[,2]=="Up",]

smc.camera <- cameraPR(treat.all$t[,9],reactome.id)
smc.camera.up <- smc.camera[smc.camera[,2]=="Up",]


nsets <- 5
all.cam <- rbind(eryth.camera.up[1:nsets,], cardioprlf.camera.up[1:nsets,],
                       cardio.camera.up[1:nsets,],endo.camera.up[1:nsets,],
                       pericyte.camera.up[1:nsets,],fibro.camera.up[1:nsets,],
                       immune.camera.up[1:nsets,],neuron.camera.up[1:nsets,],smc.camera.up[1:nsets,])

scores <- -log10(all.cam$PValue)
names(scores) <- rownames(all.cam)
names(scores) <- gsub("REACTOME_","",names(scores))
```

```{r, fig.width=12, fig.height=11}
par(mfrow=c(1,1))
par(mar=c(5,41,3,2))
barplot(scores[length(scores):1],horiz = T,las=2,col=rev(rep(c(ggplotColors(8),"grey"),each=nsets)),cex.names=0.9,
        cex.axis = 1.5,xlab="-log10(PValue)",cex.lab=1.5)
abline(v= -log10(0.05),lty=2)

```
