==> analysis/04-DE-BroadCellTypes.Rmd <==
---
title: "Differential expression analysis of broad cell type clusters"
author: "Belinda Phipson"
date: "10/25/2019"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r, echo=FALSE}
knitr::opts_chunk$set(error = TRUE)
```

# Introduction

# Load libraries and functions

```{r}
library(edgeR)
library(RColorBrewer)
library(org.Hs.eg.db)
library(limma)
library(Seurat)
library(cowplot)
library(DelayedArray)
library(scran)
library(NMF)
library(workflowr)
library(ggplot2)
library(clustree)
library(dplyr)
```

```{r}
source("code/normCounts.R")
source("code/findModes.R")
source("code/ggplotColors.R")
```

# Load the data

```{r readTargets}
targets <- read.delim("data/targets.txt",header=TRUE, stringsAsFactors = FALSE)
targets$FileName2 <- paste(targets$FileName,"/",sep="")
targets$Group_ID2 <- gsub("LV_","",targets$Group_ID)
group <- c("fetal_1","fetal_2","fetal_3",
           "non-diseased_1","non-diseased_2","non-diseased_3",
           "diseased_1","diseased_2",
           "diseased_3","diseased_4")
m <- match(group, targets$Group_ID2)
targets <- targets[m,]
```

```{rm readData}
fetal.integrated <- readRDS(file="/group/card2/Neda/MCRI_LAB/single_cell_nuclei_rnaseq/Porello-heart-snRNAseq/output/RDataObjects/fetal-int.Rds")
load(file="/group/card2/Neda/MCRI_LAB/single_cell_nuclei_rnaseq/Porello-heart-snRNAseq/output/RDataObjects/fetalObjs.Rdata")
##note: nd.integrated is also an integrated form of young heart samples. young.integrated has already been published in Sim et al., 2021
nd.integrated <- readRDS(file="/group/card2/Neda/MCRI_LAB/single_cell_nuclei_rnaseq/Porello-heart-snRNAseq/output/RDataObjects/nd-int.Rds")
load(file="/group/card2/Neda/MCRI_LAB/single_cell_nuclei_rnaseq/Porello-heart-snRNAseq/output/RDataObjects/ndObjs.Rdata")

dcm.integrated <- readRDS(file="/group/card2/Neda/MCRI_LAB/single_cell_nuclei_rnaseq/Porello-heart-snRNAseq/output/RDataObjects/dcm-int.Rds")
load(file="/group/card2/Neda/MCRI_LAB/single_cell_nuclei_rnaseq/Porello-heart-snRNAseq/output/RDataObjects/dcmObjs.Rdata")
```

```{r}
# Load unfiltered counts matrix for every sample (object all)
load("/group/card2/Neda/MCRI_LAB/single_cell_nuclei_rnaseq/Porello-heart-snRNAseq/output/RDataObjects/all-counts.Rdata")
```

# Set default clustering resolution

```{rm}
# Default 0.3
Idents(fetal.integrated) <- fetal.integrated$integrated_snn_res.0.3
DimPlot(fetal.integrated, reduction = "tsne",label=TRUE,label.size = 6)+NoLegend()

# Default 0.3
Idents(nd.integrated) <- nd.integrated$integrated_snn_res.0.3
DimPlot(nd.integrated, reduction = "tsne",label=TRUE,label.size = 6)+NoLegend()

# Default 0.3
Idents(dcm.integrated) <- dcm.integrated$integrated_snn_res.0.3
DimPlot(dcm.integrated, reduction = "tsne",label=TRUE,label.size = 6)+NoLegend()
```


# Merge all data together

```{rm}
# This data has already been generated and saved as the heart object.
heart <- merge(fetal.integrated, y = c(nd.integrated, dcm.integrated), project = "heart")

```

```{r}
heart <- readRDS("/group/card2/Neda/MCRI_LAB/must-do-projects/EnzoPorrelloLab/dilated-cardiomyopathy/data/heart-int-FND-filtered.Rds")
Idents(heart) <- heart$Broad_celltype
table(heart$orig.ident)
```

# Remove Er and CM(Prlf) cells

```{r}
heart <- subset(heart,subset = Broad_celltype != "Er")
heart <- subset(heart,subset = Broad_celltype != "CM(Prlf)")
heart$biorep <- factor(heart$biorep,levels=c("f1","f2","f3","nd1","nd2","nd3","d1","d2","d3","d4"))
table(heart$Broad_celltype,heart$biorep)
```


# Get gene annotation

```{r}
columns(org.Hs.eg.db)
DefaultAssay(heart) <- "RNA"
ann <- AnnotationDbi:::select(org.Hs.eg.db,keys=rownames(all),columns=c("SYMBOL","ENTREZID","ENSEMBL","GENENAME","CHR"),keytype = "SYMBOL")
m <- match(rownames(all),ann$SYMBOL)
ann <- ann[m,]
table(ann$SYMBOL==rownames(all))
```

```{r}
mito <- grep("mitochondrial",ann$GENENAME)
length(mito)
ribo <- grep("ribosomal",ann$GENENAME)
length(ribo)
missingEZID <- which(is.na(ann$ENTREZID))
length(missingEZID)
```

# Filter out non-informative genes

```{r markers}
m <- match(colnames(heart),colnames(all))
all.counts <- all[,m]
```

```{r}
chuck <- unique(c(mito,ribo,missingEZID))
length(chuck)
all.counts.keep <- all.counts[-chuck,]
ann.keep <- ann[-chuck,]
table(ann.keep$SYMBOL==rownames(all.counts.keep))
```

```{r}
# remove x y genes

xy <- ann.keep$CHR %in% c("X","Y")
all.counts.keep <- all.counts.keep[!xy,]
ann.keep <- ann.keep[!xy,]
```



```{r}
numzero.genes <- rowSums(all.counts.keep==0)

table(numzero.genes > (ncol(all.counts.keep)-20))

keep.genes <- numzero.genes < (ncol(all.counts.keep)-20)
table(keep.genes)

all.keep <- all.counts.keep[keep.genes,]
dim(all.keep)

ann.keep <- ann.keep[keep.genes,]

table(colnames(heart)==colnames(all.keep))

table(rownames(all.keep)==ann.keep$SYMBOL)
```

# Create pseudobulk samples

```{r}
broadct <- factor(heart$Broad_celltype)
sam <- factor(heart$biorep,levels=c("f1","f2","f3","nd1","nd2","nd3","d1","d2","d3","d4"))
newgrp <- paste(broadct,sam,sep=".")
newgrp <- factor(newgrp,levels=paste(rep(levels(broadct),each=10),levels(sam),sep="."))
table(newgrp)
```


```{r pb}
des <- model.matrix(~0+newgrp)
colnames(des) <- levels(newgrp)

dim(des)

pb <- all.keep %*% des
```

```{r dgelist}
y.pb <- DGEList(pb)
saminfo <- matrix(unlist(strsplit(colnames(y.pb$counts),split="[.]")),ncol=2,byrow=TRUE)
bct <- factor(saminfo[,1])
indiv <- factor(saminfo[,2])
group <- rep(NA,ncol(y.pb))
group[grep("f",indiv)] <- "Fetal"
group[grep("nd",indiv)] <- "ND"
group[grep("^d",indiv)] <- "DCM"
group <- factor(group,levels=c("Fetal","ND","DCM"))
targets.pb <- data.frame(bct,indiv,group)
targets.pb$Sex <- c("Male","Male","Female","Male","Female","Male","Male","Female","Female","Female")
```

```{r}
par(mfrow=c(1,1))
sexgroup <- paste(group,targets.pb$Sex,sep=".")
plotMDS.DGEList(y.pb,gene.selection="common",col=ggplotColors(7)[bct],pch=c(0,15,1,16,2,17)[factor(sexgroup)],cex=2)
legend("bottomright",legend=levels(bct),fill=ggplotColors(7),bty="n")
legend("bottomleft",legend=c("Female","Male"),pch=c(1,16),cex=1.5)
legend("topright",legend=c("Fetal","ND","DCM"),pch=c(1,2,0),cex=1.5)

plotMDS.DGEList(y.pb,gene.selection="common",col=ggplotColors(7)[bct],pch=c(0,15,1,16,2,17)[factor(sexgroup)],cex=2,dim=c(3,4))
legend("bottomright",legend=levels(bct),fill=ggplotColors(7),bty="n")
legend("top",legend=c("Female","Male"),pch=c(1,16),cex=1.5)
legend("topright",legend=c("Fetal","ND","DCM"),pch=c(1,2,0),cex=1.5)

plotMDS.DGEList(y.pb,gene.selection="common",col=ggplotColors(7)[bct],pch=c(0,15,1,16,2,17)[factor(sexgroup)],cex=2,dim=c(5,6))
legend("bottomleft",legend=levels(bct),fill=ggplotColors(7),bty="n")
legend("top",legend=c("Female","Male"),pch=c(1,16),cex=1.5)
legend("bottomright",legend=c("Fetal","ND","DCM"),pch=c(1,2,0),cex=1.5)
```


# Differential expression analysis

```{r}
table(rownames(y.pb)==ann.keep$SYMBOL)
y.pb$genes <- ann.keep

bct2 <- as.character(bct)
bct2[bct2 == "CM"] <- "cardio"
bct2[bct2 == "Endo"] <- "endo"
bct2[bct2 == "Pericyte"] <- "peri"
bct2[bct2 == "Fib"] <- "fibro"
bct2[bct2 == "Immune"] <- "immune"
bct2[bct2 == "Neuron"] <- "neurons"
bct2[bct2 == "Smc"] <- "smc"

newgrp <- paste(bct2,group,targets.pb$Sex,sep=".")
newgrp <- factor(newgrp)

design <- model.matrix(~0+newgrp)
colnames(design) <- levels(newgrp)

v <- voom(y.pb,design,plot=TRUE,normalize.method = "cyclicloess")

fit <- lmFit(v,design)

```

## Cardiomyocytes

```{r}
cont.cardio <- makeContrasts(NvF = 0.5*(cardio.ND.Male + cardio.ND.Female) - 0.5*(cardio.Fetal.Male + cardio.Fetal.Female),
                         DvF = 0.5*(cardio.DCM.Male + cardio.DCM.Female) - 0.5*(cardio.Fetal.Male + cardio.Fetal.Female),
                         DvN = 0.5*(cardio.DCM.Male + cardio.DCM.Female) - 0.5*(cardio.ND.Male + cardio.ND.Female),
                         SexFetal = cardio.Fetal.Male - cardio.Fetal.Female,
                         SexND = cardio.ND.Male - cardio.ND.Female,
                         SexDCM = cardio.DCM.Male - cardio.DCM.Female,
                         InteractionDF = (cardio.DCM.Male - cardio.Fetal.Male) - (cardio.DCM.Female - cardio.Fetal.Female),
                         InteractionNF = (cardio.ND.Male - cardio.Fetal.Male) - (cardio.ND.Female - cardio.Fetal.Female),
                         InteractionDN = (cardio.DCM.Male - cardio.ND.Male) - (cardio.DCM.Female - cardio.ND.Female),
                         levels=design)
fit.cardio <- contrasts.fit(fit,contrasts = cont.cardio)
fit.cardio <- eBayes(fit.cardio,robust=TRUE)

summary(decideTests(fit.cardio))

treat.cardio <- treat(fit.cardio,lfc=0.5)

dt.cardio<-decideTests(treat.cardio)

summary(dt.cardio)
```

```{r}
par(mfrow=c(3,3))
for(i in 1:9){
  plotMD(treat.cardio,coef=i,status=dt.cardio[,i],hl.col=c("blue","red"))
  abline(h=0,col="grey")
}
```


```{r}
par(mfrow=c(1,1))
par(mar=c(7,4,2,2))
barplot(summary(dt.cardio)[-2,],beside=TRUE,legend=TRUE,col=c("blue","red"),ylab="Number of significant genes",las=2,xlab="")
title("Cardiomyocytes")
```

### ND vs Fetal
```{r}
options(digits=3)
topTreat(treat.cardio,coef=1,n=20,p.value=0.05)[,-c(1:3)]
```

### DCM vs Fetal
```{r}
topTreat(treat.cardio,coef=2,n=20,p.value=0.05)[,-c(1:3)]
```

### DCM vs ND
```{r}
topTreat(treat.cardio,coef=3,n=20,p.value=0.05)[,-c(1:3)]
```

### Male vs Female in Fetal samples
```{r}
topTreat(treat.cardio,coef=4,n=20,p.value=0.05)[,-c(1:3)]
```

### Male vs Female in ND samples
```{r}
topTreat(treat.cardio,coef=5,n=20,p.value=0.05)[,-c(1:3)]
```

### Male vs Female in DCM samples
```{r}
topTreat(treat.cardio,coef=6,n=20,p.value=0.05)[,-c(1:3)]
```

### Interaction of sex differences between DCM and Fetal
```{r}
topTreat(treat.cardio,coef=7,n=20,p.value=0.05)[,-c(1:3)]
```

### Interaction of sex differences between ND and Fetal
```{r}
topTreat(treat.cardio,coef=8,n=20,p.value=0.05)[,-c(1:3)]
```

### Interaction of sex differences between DCM and ND
```{r}
topTreat(treat.cardio,coef=9,n=20,p.value=0.05)[,-c(1:3)]
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.cardio$coefficients[,1],treat.cardio$coefficients[,2],xlab="logFC: ND vs Fetal", ylab="logFC: DCM vs Fetal",main="Cardiomyocytes",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0,col=colours()[c(175)])
sig.genes <-  dt.cardio[,1] !=0 | dt.cardio[,2] != 0
text(treat.cardio$coefficients[sig.genes,1],treat.cardio$coefficients[sig.genes,2],labels=rownames(dt.cardio)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.cardio$coefficients[,4],treat.cardio$coefficients[,5],xlab="logFC: Fetal Male vs Female", ylab="logFC: ND Male vs Female",main="Cardiomyocytes: Sex differences - Fetal and ND",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.cardio[,4] !=0 | dt.cardio[,5] != 0
text(treat.cardio$coefficients[sig.genes,4],treat.cardio$coefficients[sig.genes,5],labels=rownames(dt.cardio)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.cardio$coefficients[,4],treat.cardio$coefficients[,6],xlab="logFC: Fetal Male vs Female", ylab="logFC: DCM Male vs Female",main="Cardiomyocytes: Sex differences - Fetal and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.cardio[,4] !=0 | dt.cardio[,6] != 0
text(treat.cardio$coefficients[sig.genes,4],treat.cardio$coefficients[sig.genes,6],labels=rownames(dt.cardio)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.cardio$coefficients[,5],treat.cardio$coefficients[,6],xlab="logFC: ND Male vs Female", ylab="logFC: DCM Male vs Female",main="Cardiomyocytes: Sex differences - ND and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
sig.genes <- dt.cardio[,5] !=0 | dt.cardio[,6] != 0
text(treat.cardio$coefficients[sig.genes,5],treat.cardio$coefficients[sig.genes,6],labels=rownames(dt.cardio)[sig.genes],col=2,cex=0.6)
abline(h=0,v=0)
```


```{r}
par(mar=c(5,5,3,2))
plot(treat.cardio$coefficients[,4],treat.cardio$coefficients[,5],xlab="logFC: Fetal Male vs Female", ylab="logFC: ND Male vs Female",main="Cardiomyocytes: Interaction term: Fetal vs ND",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.cardio[,8] !=0 
text(treat.cardio$coefficients[sig.genes,4],treat.cardio$coefficients[sig.genes,5],labels=rownames(dt.cardio)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.cardio$coefficients[,4],treat.cardio$coefficients[,6],xlab="logFC: Fetal Male vs Female", ylab="logFC: DCM Male vs Female",main="Cardiomyocytes: Interaction term: Fetal and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.cardio[,7] !=0 
text(treat.cardio$coefficients[sig.genes,4],treat.cardio$coefficients[sig.genes,6],labels=rownames(dt.cardio)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.cardio$coefficients[,5],treat.cardio$coefficients[,6],xlab="logFC: ND Male vs Female", ylab="logFC: DCM Male vs Female",main="Cardiomyocytes: Interaction term: ND and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.cardio[,9] !=0 
text(treat.cardio$coefficients[sig.genes,5],treat.cardio$coefficients[sig.genes,6],labels=rownames(dt.cardio)[sig.genes],col=2,cex=0.6)
```

## Pericytes

```{r}
cont.peri <- makeContrasts(NvF = 0.5*(peri.ND.Male + peri.ND.Female) - 0.5*(peri.Fetal.Male + peri.Fetal.Female),
                         DvF = 0.5*(peri.DCM.Male + peri.DCM.Female) - 0.5*(peri.Fetal.Male + peri.Fetal.Female),
                         DvN = 0.5*(peri.DCM.Male + peri.DCM.Female) - 0.5*(peri.ND.Male + peri.ND.Female),
                         SexFetal = peri.Fetal.Male - peri.Fetal.Female,
                         SexND = peri.ND.Male - peri.ND.Female,
                         SexDCM = peri.DCM.Male - peri.DCM.Female,
                         InteractionDF = (peri.DCM.Male - peri.Fetal.Male) - (peri.DCM.Female - peri.Fetal.Female),
                         InteractionNF = (peri.ND.Male - peri.Fetal.Male) - (peri.ND.Female - peri.Fetal.Female),
                         InteractionDN = (peri.DCM.Male - peri.ND.Male) - (peri.DCM.Female - peri.ND.Female),
                         levels=design)
fit.peri <- contrasts.fit(fit,contrasts = cont.peri)
fit.peri <- eBayes(fit.peri,robust=TRUE)

summary(decideTests(fit.peri))

treat.peri <- treat(fit.peri,lfc=0.5)

dt.peri<-decideTests(treat.peri)

summary(dt.peri)
```

```{r}
par(mfrow=c(3,3))
for(i in 1:9){
  plotMD(treat.peri,coef=i,status=dt.peri[,i],hl.col=c("blue","red"))
  abline(h=0,col="grey")
}
```


```{r}
par(mfrow=c(1,1))
par(mar=c(7,4,2,2))
barplot(summary(dt.peri)[-2,],beside=TRUE,legend=TRUE,col=c("blue","red"),ylab="Number of significant genes",las=2)
title("Pericytes")
```

### ND vs Fetal
```{r}
options(digits=3)
topTreat(treat.peri,coef=1,n=20,p.value=0.05)[,-c(1:3)]
```

### DCM vs Fetal
```{r}
topTreat(treat.peri,coef=2,n=20,p.value=0.05)[,-c(1:3)]
```

### DCM vs ND
```{r}
topTreat(treat.peri,coef=3)[,-c(1:3)]
```

### Male vs Female in Fetal samples
```{r}
topTreat(treat.peri,coef=4)[,-c(1:3)]
```

### Male vs Female in ND samples
```{r}
topTreat(treat.peri,coef=5)[,-c(1:3)]
```

### Male vs Female in DCM samples
```{r}
topTreat(treat.peri,coef=6)[,-c(1:3)]
```

### Interaction of sex differences between DCM and Fetal
```{r}
topTreat(treat.peri,coef=7)[,-c(1:3)]
```

### Interaction of sex differences between ND and Fetal
```{r}
topTreat(treat.peri,coef=8)[,-c(1:3)]
```

### Interaction of sex differences between DCM and ND
```{r}
topTreat(treat.peri,coef=9)[,-c(1:3)]
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.peri$coefficients[,1],treat.peri$coefficients[,2],xlab="logFC: ND vs Fetal", ylab="logFC: DCM vs Fetal",main="Pericytes",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0,col=colours()[c(175)])
sig.genes <-  dt.peri[,1] !=0 | dt.peri[,2] != 0
text(treat.peri$coefficients[sig.genes,1],treat.peri$coefficients[sig.genes,2],labels=rownames(dt.peri)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.peri$coefficients[,4],treat.peri$coefficients[,5],xlab="logFC: Fetal Male vs Female", ylab="logFC: ND Male vs Female",main="Pericyte: Sex differences - Fetal and ND",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.peri[,4] !=0 | dt.peri[,5] != 0
text(treat.peri$coefficients[sig.genes,4],treat.peri$coefficients[sig.genes,5],labels=rownames(dt.peri)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.peri$coefficients[,4],treat.peri$coefficients[,6],xlab="logFC: Fetal Male vs Female", ylab="logFC: DCM Male vs Female",main="Pericyte: Sex differences - Fetal and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.peri[,4] !=0 | dt.peri[,6] != 0
text(treat.peri$coefficients[sig.genes,4],treat.peri$coefficients[sig.genes,6],labels=rownames(dt.peri)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.peri$coefficients[,5],treat.peri$coefficients[,6],xlab="logFC: ND Male vs Female", ylab="logFC: DCM Male vs Female",main="Pericyte: Sex differences - ND and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
sig.genes <- dt.peri[,5] !=0 | dt.peri[,6] != 0
text(treat.peri$coefficients[sig.genes,5],treat.peri$coefficients[sig.genes,6],labels=rownames(dt.peri)[sig.genes],col=2,cex=0.6)
abline(h=0,v=0)
```


```{r}
par(mar=c(5,5,3,2))
plot(treat.peri$coefficients[,4],treat.peri$coefficients[,5],xlab="logFC: Fetal Male vs Female", ylab="logFC: ND Male vs Female",main="Pericyte: Interaction term: Fetal vs ND",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.peri[,8] !=0 
if(sum(sig.genes)!=0){
  text(treat.peri$coefficients[sig.genes,4],treat.peri$coefficients[sig.genes,5],
       labels=rownames(dt.peri)[sig.genes],col=2,cex=0.6)
}
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.peri$coefficients[,4],treat.peri$coefficients[,6],xlab="logFC: Fetal Male vs Female", ylab="logFC: DCM Male vs Female",main="Pericyte: Interaction term: Fetal and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.peri[,7] !=0 
if(sum(sig.genes)!=0){
  text(treat.peri$coefficients[sig.genes,4],treat.peri$coefficients[sig.genes,6],
       labels=rownames(dt.peri)[sig.genes],col=2,cex=0.6)
}
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.peri$coefficients[,5],treat.peri$coefficients[,6],xlab="logFC: ND Male vs Female", ylab="logFC: DCM Male vs Female",main="Pericyte: Interaction term: ND and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.peri[,9] !=0 
if(sum(sig.genes)!=0){
  text(treat.peri$coefficients[sig.genes,5],treat.peri$coefficients[sig.genes,6],
       labels=rownames(dt.peri)[sig.genes],col=2,cex=0.6)
}
```

## Fibroblasts

```{r}
cont.fibro <- makeContrasts(NvF = 0.5*(fibro.ND.Male + fibro.ND.Female) - 0.5*(fibro.Fetal.Male + fibro.Fetal.Female),
                         DvF = 0.5*(fibro.DCM.Male + fibro.DCM.Female) - 0.5*(fibro.Fetal.Male + fibro.Fetal.Female),
                         DvN = 0.5*(fibro.DCM.Male + fibro.DCM.Female) - 0.5*(fibro.ND.Male + fibro.ND.Female),
                         SexFetal = fibro.Fetal.Male - fibro.Fetal.Female,
                         SexND = fibro.ND.Male - fibro.ND.Female,
                         SexDCM = fibro.DCM.Male - fibro.DCM.Female,
                         InteractionDF = (fibro.DCM.Male - fibro.Fetal.Male) - (fibro.DCM.Female - fibro.Fetal.Female),
                         InteractionNF = (fibro.ND.Male - fibro.Fetal.Male) - (fibro.ND.Female - fibro.Fetal.Female),
                         InteractionDN = (fibro.DCM.Male - fibro.ND.Male) - (fibro.DCM.Female - fibro.ND.Female),
                         levels=design)
fit.fibro <- contrasts.fit(fit,contrasts = cont.fibro)
fit.fibro <- eBayes(fit.fibro,robust=TRUE)

summary(decideTests(fit.fibro))

treat.fibro <- treat(fit.fibro,lfc=0.5)

dt.fibro<-decideTests(treat.fibro)

summary(dt.fibro)
```

```{r}
par(mfrow=c(3,3))
for(i in 1:9){
  plotMD(treat.fibro,coef=i,status=dt.fibro[,i],hl.col=c("blue","red"))
  abline(h=0,col="grey")
}
```


```{r}
par(mfrow=c(1,1))
par(mar=c(7,4,2,2))
barplot(summary(dt.fibro)[-2,],beside=TRUE,legend=TRUE,col=c("blue","red"),ylab="Number of significant genes",las=2)
title("fibroblast cells")
```

### ND vs Fetal
```{r}
options(digits=3)
topTreat(treat.fibro,coef=1,n=20,p.value=0.05)[,-c(1:3)]
```

### DCM vs Fetal
```{r}
topTreat(treat.fibro,coef=2,n=20,p.value=0.05)[,-c(1:3)]
```

### DCM vs ND
```{r}
topTreat(treat.fibro,coef=3,n=20,p.value=0.05)[,-c(1:3)]
```

### Male vs Female in Fetal samples
```{r}
topTreat(treat.fibro,coef=4)[,-c(1:3)]
```

### Male vs Female in ND samples
```{r}
topTreat(treat.fibro,coef=5)[,-c(1:3)]
```

### Male vs Female in DCM samples
```{r}
topTreat(treat.fibro,coef=6)[,-c(1:3)]
```

### Interaction of sex differences between DCM and Fetal
```{r}
topTreat(treat.fibro,coef=7)[,-c(1:3)]
```

### Interaction of sex differences between ND and Fetal
```{r}
topTreat(treat.fibro,coef=8)[,-c(1:3)]
```

### Interaction of sex differences between DCM and ND
```{r}
topTreat(treat.fibro,coef=9)[,-c(1:3)]
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.fibro$coefficients[,1],treat.fibro$coefficients[,2],xlab="logFC: ND vs Fetal", ylab="logFC: DCM vs Fetal",main="Fibroblasts",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0,col=colours()[c(175)])
sig.genes <-  dt.fibro[,1] !=0 | dt.fibro[,2] != 0
text(treat.fibro$coefficients[sig.genes,1],treat.fibro$coefficients[sig.genes,2],labels=rownames(dt.fibro)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.fibro$coefficients[,4],treat.fibro$coefficients[,5],xlab="logFC: Fetal Male vs Female", ylab="logFC: ND Male vs Female",main="Fibroblasts: Sex differences - Fetal and ND",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.fibro[,4] !=0 | dt.fibro[,5] != 0
text(treat.fibro$coefficients[sig.genes,4],treat.fibro$coefficients[sig.genes,5],labels=rownames(dt.fibro)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.fibro$coefficients[,4],treat.fibro$coefficients[,6],xlab="logFC: Fetal Male vs Female", ylab="logFC: DCM Male vs Female",main="Fibroblasts: Sex differences - Fetal and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.fibro[,4] !=0 | dt.fibro[,6] != 0
text(treat.fibro$coefficients[sig.genes,4],treat.fibro$coefficients[sig.genes,6],labels=rownames(dt.fibro)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.fibro$coefficients[,5],treat.fibro$coefficients[,6],xlab="logFC: ND Male vs Female", ylab="logFC: DCM Male vs Female",main="Fibroblasts: Sex differences - ND and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
sig.genes <- dt.fibro[,5] !=0 | dt.fibro[,6] != 0
text(treat.fibro$coefficients[sig.genes,5],treat.fibro$coefficients[sig.genes,6],labels=rownames(dt.fibro)[sig.genes],col=2,cex=0.6)
abline(h=0,v=0)
```


```{r}
par(mar=c(5,5,3,2))
plot(treat.fibro$coefficients[,4],treat.fibro$coefficients[,5],xlab="logFC: Fetal Male vs Female", ylab="logFC: ND Male vs Female",main="Fibroblasts: Interaction term: Fetal vs ND",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.fibro[,8] !=0 
if(sum(sig.genes)!=0){
  text(treat.fibro$coefficients[sig.genes,4],treat.fibro$coefficients[sig.genes,5],
       labels=rownames(dt.fibro)[sig.genes],col=2,cex=0.6)
}
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.fibro$coefficients[,4],treat.fibro$coefficients[,6],xlab="logFC: Fetal Male vs Female", ylab="logFC: DCM Male vs Female",main="Fibroblast: Interaction term: Fetal and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.fibro[,7] !=0 
text(treat.fibro$coefficients[sig.genes,4],treat.fibro$coefficients[sig.genes,6],labels=rownames(dt.fibro)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.fibro$coefficients[,5],treat.fibro$coefficients[,6],xlab="logFC: ND Male vs Female", ylab="logFC: DCM Male vs Female",main="Fibroblasts: Interaction term: ND and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.fibro[,9] !=0 
text(treat.fibro$coefficients[sig.genes,5],treat.fibro$coefficients[sig.genes,6],labels=rownames(dt.fibro)[sig.genes],col=2,cex=0.6)
```


## Smooth muscle cells

```{r}
cont.smc <- makeContrasts(NvF = 0.5*(smc.ND.Male + smc.ND.Female) - 0.5*(smc.Fetal.Male + smc.Fetal.Female),
                         DvF = 0.5*(smc.DCM.Male + smc.DCM.Female) - 0.5*(smc.Fetal.Male + smc.Fetal.Female),
                         DvN = 0.5*(smc.DCM.Male + smc.DCM.Female) - 0.5*(smc.ND.Male + smc.ND.Female),
                         SexFetal = smc.Fetal.Male - smc.Fetal.Female,
                         SexND = smc.ND.Male - smc.ND.Female,
                         SexDCM = smc.DCM.Male - smc.DCM.Female,
                         InteractionDF = (smc.DCM.Male - smc.Fetal.Male) - (smc.DCM.Female - smc.Fetal.Female),
                         InteractionNF = (smc.ND.Male - smc.Fetal.Male) - (smc.ND.Female - smc.Fetal.Female),
                         InteractionDN = (smc.DCM.Male - smc.ND.Male) - (smc.DCM.Female - smc.ND.Female),
                         levels=design)
fit.smc <- contrasts.fit(fit,contrasts = cont.smc)
fit.smc <- eBayes(fit.smc,robust=TRUE)

summary(decideTests(fit.smc))

treat.smc <- treat(fit.smc,lfc=0.5)

dt.smc<-decideTests(treat.smc)

summary(dt.smc)
```

```{r}
par(mfrow=c(3,3))
for(i in 1:9){
  plotMD(treat.smc,coef=i,status=dt.smc[,i],hl.col=c("blue","red"))
  abline(h=0,col="grey")
}
```


```{r}
par(mfrow=c(1,1))
par(mar=c(7,4,2,2))
barplot(summary(dt.smc)[-2,],beside=TRUE,legend=TRUE,col=c("blue","red"),ylab="Number of significant genes",las=2)
title("smooth muscle cells")
```

### ND vs Fetal
```{r}
options(digits=3)
topTreat(treat.smc,coef=1,n=20,p.value=0.05)[,-c(1:3)]
```

### DCM vs Fetal
```{r}
topTreat(treat.smc,coef=2,n=20,p.value=0.05)[,-c(1:3)]
```

### DCM vs ND
```{r}
topTreat(treat.smc,coef=3)[,-c(1:3)]
```

### Male vs Female in Fetal samples
```{r}
topTreat(treat.smc,coef=4,n=20,p.value=0.05)[,-c(1:3)]
```

### Male vs Female in ND samples
```{r}
topTreat(treat.smc,coef=5)[,-c(1:3)]
```

### Male vs Female in DCM samples
```{r}
topTreat(treat.smc,coef=6)[,-c(1:3)]
```

### Interaction of sex differences between DCM and Fetal
```{r}
topTreat(treat.smc,coef=7)[,-c(1:3)]
```

### Interaction of sex differences between ND and Fetal
```{r}
topTreat(treat.smc,coef=8)[,-c(1:3)]
```

### Interaction of sex differences between DCM and ND
```{r}
topTreat(treat.smc,coef=9)[,-c(1:3)]
```

```{r}
par(mar=c(5,5,3,2))
par(mar=c(5,5,3,2))
plot(treat.smc$coefficients[,1],treat.smc$coefficients[,2],xlab="logFC: ND vs Fetal", ylab="logFC: DCM vs Fetal",main="smooth muscle cells",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0,col=colours()[c(175)])
sig.genes <-  dt.smc[,1] !=0 | dt.smc[,2] != 0
text(treat.smc$coefficients[sig.genes,1],treat.smc$coefficients[sig.genes,2],labels=rownames(dt.smc)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.smc$coefficients[,4],treat.smc$coefficients[,5],xlab="logFC: Fetal Male vs Female", ylab="logFC: ND Male vs Female",main="smooth muscle cells: Sex differences - Fetal and ND",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.smc[,4] !=0 | dt.smc[,5] != 0
text(treat.smc$coefficients[sig.genes,4],treat.smc$coefficients[sig.genes,5],labels=rownames(dt.smc)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.smc$coefficients[,4],treat.smc$coefficients[,6],xlab="logFC: Fetal Male vs Female", ylab="logFC: DCM Male vs Female",main="smooth muscle cells: Sex differences - Fetal and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.smc[,4] !=0 | dt.smc[,6] != 0
text(treat.smc$coefficients[sig.genes,4],treat.smc$coefficients[sig.genes,6],labels=rownames(dt.smc)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.smc$coefficients[,5],treat.smc$coefficients[,6],xlab="logFC: ND Male vs Female", ylab="logFC: DCM Male vs Female",main="smooth muscle cells: Sex differences - ND and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
sig.genes <- dt.smc[,5] !=0 | dt.smc[,6] != 0
text(treat.smc$coefficients[sig.genes,5],treat.smc$coefficients[sig.genes,6],labels=rownames(dt.smc)[sig.genes],col=2,cex=0.6)
abline(h=0,v=0)
```


```{r}
par(mar=c(5,5,3,2))
plot(treat.smc$coefficients[,4],treat.smc$coefficients[,5],xlab="logFC: Fetal Male vs Female", ylab="logFC: ND Male vs Female",main="smooth muscle cells: Interaction term: Fetal vs ND",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.smc[,8] !=0 
text(treat.smc$coefficients[sig.genes,4],treat.smc$coefficients[sig.genes,5],labels=rownames(dt.smc)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.smc$coefficients[,4],treat.smc$coefficients[,6],xlab="logFC: Fetal Male vs Female", ylab="logFC: DCM Male vs Female",main="smooth muscle cells: Interaction term: Fetal and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.smc[,7] !=0 
text(treat.smc$coefficients[sig.genes,4],treat.smc$coefficients[sig.genes,6],labels=rownames(dt.smc)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.smc$coefficients[,5],treat.smc$coefficients[,6],xlab="logFC: ND Male vs Female", ylab="logFC: DCM Male vs Female",main="smooth muscle cells: Interaction term: ND and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.smc[,9] !=0 
if(sum(sig.genes)!=0){
  text(treat.smc$coefficients[sig.genes,5],treat.smc$coefficients[sig.genes,6],
       labels=rownames(dt.smc)[sig.genes],col=2,cex=0.6)
}
```


## Endothelial cells

```{r}
cont.endo <- makeContrasts(NvF = 0.5*(endo.ND.Male + endo.ND.Female) - 0.5*(endo.Fetal.Male + endo.Fetal.Female),
                         DvF = 0.5*(endo.DCM.Male + endo.DCM.Female) - 0.5*(endo.Fetal.Male + endo.Fetal.Female),
                         DvN = 0.5*(endo.DCM.Male + endo.DCM.Female) - 0.5*(endo.ND.Male + endo.ND.Female),
                         SexFetal = endo.Fetal.Male - endo.Fetal.Female,
                         SexND = endo.ND.Male - endo.ND.Female,
                         SexDCM = endo.DCM.Male - endo.DCM.Female,
                         InteractionDF = (endo.DCM.Male - endo.Fetal.Male) - (endo.DCM.Female - endo.Fetal.Female),
                         InteractionNF = (endo.ND.Male - endo.Fetal.Male) - (endo.ND.Female - endo.Fetal.Female),
                         InteractionDN = (endo.DCM.Male - endo.ND.Male) - (endo.DCM.Female - endo.ND.Female),
                         levels=design)
fit.endo <- contrasts.fit(fit,contrasts = cont.endo)
fit.endo <- eBayes(fit.endo,robust=TRUE)

summary(decideTests(fit.endo))

treat.endo <- treat(fit.endo,lfc=0.5)

dt.endo<-decideTests(treat.endo)

summary(dt.endo)
```

```{r}
par(mfrow=c(3,3))
for(i in 1:9){
  plotMD(treat.endo,coef=i,status=dt.endo[,i],hl.col=c("blue","red"))
  abline(h=0,col="grey")
}
```


```{r}
par(mfrow=c(1,1))
par(mar=c(7,4,2,2))
barplot(summary(dt.endo)[-2,],beside=TRUE,legend=TRUE,col=c("blue","red"),ylab="Number of significant genes",las=2)
title("Endothelial cells")
```

### ND vs Fetal
```{r}
options(digits=3)
topTreat(treat.endo,coef=1,n=20,p.value=0.05)[,-c(1:3)]
```

### DCM vs Fetal
```{r}
topTreat(treat.endo,coef=2,n=20,p.value=0.05)[,-c(1:3)]
```

### DCM vs ND
```{r}
topTreat(treat.endo,coef=3)[,-c(1:3)]
```

### Male vs Female in Fetal samples
```{r}
topTreat(treat.endo,coef=4)[,-c(1:3)]
```

### Male vs Female in ND samples
```{r}
topTreat(treat.endo,coef=5)[,-c(1:3)]
```

### Male vs Female in DCM samples
```{r}
topTreat(treat.endo,coef=6)[,-c(1:3)]
```

### Interaction of sex differences between DCM and Fetal
```{r}
topTreat(treat.endo,coef=7)[,-c(1:3)]
```

### Interaction of sex differences between ND and Fetal
```{r}
topTreat(treat.endo,coef=8)[,-c(1:3)]
```

### Interaction of sex differences between DCM and ND
```{r}
topTreat(treat.endo,coef=9)[,-c(1:3)]
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.endo$coefficients[,1],treat.endo$coefficients[,2],xlab="logFC: ND vs Fetal", ylab="logFC: DCM vs Fetal",main="endothelial cells",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0,col=colours()[c(175)])
sig.genes <-  dt.endo[,1] !=0 | dt.endo[,2] != 0
text(treat.endo$coefficients[sig.genes,1],treat.endo$coefficients[sig.genes,2],labels=rownames(dt.endo)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.endo$coefficients[,4],treat.endo$coefficients[,5],xlab="logFC: Fetal Male vs Female", ylab="logFC: ND Male vs Female",main="endothelial cells: Sex differences - Fetal and ND",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.endo[,4] !=0 | dt.endo[,5] != 0
text(treat.endo$coefficients[sig.genes,4],treat.endo$coefficients[sig.genes,5],labels=rownames(dt.endo)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.endo$coefficients[,4],treat.endo$coefficients[,6],xlab="logFC: Fetal Male vs Female", ylab="logFC: DCM Male vs Female",main="endothelial cells: Sex differences - Fetal and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.endo[,4] !=0 | dt.endo[,6] != 0
text(treat.endo$coefficients[sig.genes,4],treat.endo$coefficients[sig.genes,6],labels=rownames(dt.endo)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.endo$coefficients[,5],treat.endo$coefficients[,6],xlab="logFC: ND Male vs Female", ylab="logFC: DCM Male vs Female",main="endothelial cells: Sex differences - ND and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
sig.genes <- dt.endo[,5] !=0 | dt.endo[,6] != 0
text(treat.endo$coefficients[sig.genes,5],treat.endo$coefficients[sig.genes,6],labels=rownames(dt.endo)[sig.genes],col=2,cex=0.6)
abline(h=0,v=0)
```


```{r}
par(mar=c(5,5,3,2))
plot(treat.endo$coefficients[,4],treat.endo$coefficients[,5],xlab="logFC: Fetal Male vs Female", ylab="logFC: ND Male vs Female",main="endothelial cells: Interaction term: Fetal vs ND",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.endo[,8] !=0
if(length(rownames(dt.endo)[sig.genes]) !=0)
text(treat.endo$coefficients[sig.genes,4],treat.endo$coefficients[sig.genes,5],labels=rownames(dt.endo)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.endo$coefficients[,4],treat.endo$coefficients[,6],xlab="logFC: Fetal Male vs Female", ylab="logFC: DCM Male vs Female",main="endothelial cells: Interaction term: Fetal and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.endo[,7] !=0 
if(length(rownames(dt.endo)[sig.genes]) !=0)
text(treat.endo$coefficients[sig.genes,4],treat.endo$coefficients[sig.genes,6],labels=rownames(dt.endo)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.endo$coefficients[,5],treat.endo$coefficients[,6],xlab="logFC: ND Male vs Female", ylab="logFC: DCM Male vs Female",main="endothelial cells: Interaction term: ND and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.endo[,9] !=0 
if(length(rownames(dt.endo)[sig.genes]) !=0)
text(treat.endo$coefficients[sig.genes,5],treat.endo$coefficients[sig.genes,6],labels=rownames(dt.endo)[sig.genes],col=2,cex=0.6)
```

## Immune cells

```{r}
cont.immune <- makeContrasts(NvF = 0.5*(immune.ND.Male + immune.ND.Female) - 0.5*(immune.Fetal.Male + immune.Fetal.Female),
                         DvF = 0.5*(immune.DCM.Male + immune.DCM.Female) - 0.5*(immune.Fetal.Male + immune.Fetal.Female),
                         DvN = 0.5*(immune.DCM.Male + immune.DCM.Female) - 0.5*(immune.ND.Male + immune.ND.Female),
                         SexFetal = immune.Fetal.Male - immune.Fetal.Female,
                         SexND = immune.ND.Male - immune.ND.Female,
                         SexDCM = immune.DCM.Male - immune.DCM.Female,
                         InteractionDF = (immune.DCM.Male - immune.Fetal.Male) - (immune.DCM.Female - immune.Fetal.Female),
                         InteractionNF = (immune.ND.Male - immune.Fetal.Male) - (immune.ND.Female - immune.Fetal.Female),
                         InteractionDN = (immune.DCM.Male - immune.ND.Male) - (immune.DCM.Female - immune.ND.Female),
                         levels=design)
fit.immune <- contrasts.fit(fit,contrasts = cont.immune)
fit.immune <- eBayes(fit.immune,robust=TRUE)

summary(decideTests(fit.immune))

treat.immune <- treat(fit.immune,lfc=0.5)

dt.immune<-decideTests(treat.immune)

summary(dt.immune)
```

```{r}
par(mfrow=c(3,3))
for(i in 1:9){
  plotMD(treat.immune,coef=i,status=dt.immune[,i],hl.col=c("blue","red"))
  abline(h=0,col="grey")
}
```


```{r}
par(mfrow=c(1,1))
par(mar=c(7,4,2,2))
barplot(summary(dt.immune)[-2,],beside=TRUE,legend=TRUE,col=c("blue","red"),ylab="Number of significant genes",las=2)
title("Immune cells")
```

### ND vs Fetal
```{r}
options(digits=3)
topTreat(treat.immune,coef=1,n=20,p.value=0.05)[,-c(1:3)]
```

### DCM vs Fetal
```{r}
topTreat(treat.immune,coef=2,n=20,p.value=0.05)[,-c(1:3)]
```

### DCM vs ND
```{r}
topTreat(treat.immune,coef=3)[,-c(1:3)]
```

### Male vs Female in Fetal samples
```{r}
topTreat(treat.immune,coef=4)[,-c(1:3)]
```

### Male vs Female in ND samples
```{r}
topTreat(treat.immune,coef=5)[,-c(1:3)]
```

### Male vs Female in DCM samples
```{r}
topTreat(treat.immune,coef=6)[,-c(1:3)]
```

### Interaction of sex differences between DCM and Fetal
```{r}
topTreat(treat.immune,coef=7)[,-c(1:3)]
```

### Interaction of sex differences between ND and Fetal
```{r}
topTreat(treat.immune,coef=8)[,-c(1:3)]
```

### Interaction of sex differences between DCM and ND
```{r}
topTreat(treat.immune,coef=9)[,-c(1:3)]
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.immune$coefficients[,1],treat.immune$coefficients[,2],xlab="logFC: ND vs Fetal", ylab="logFC: DCM vs Fetal",main="immune cells",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0,col=colours()[c(175)])
sig.genes <-  dt.immune[,1] !=0 | dt.immune[,2] != 0
text(treat.immune$coefficients[sig.genes,1],treat.immune$coefficients[sig.genes,2],labels=rownames(dt.immune)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.immune$coefficients[,4],treat.immune$coefficients[,5],xlab="logFC: Fetal Male vs Female", ylab="logFC: ND Male vs Female",main="immune cells: Sex differences - Fetal and ND",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.immune[,4] !=0 | dt.immune[,5] != 0
text(treat.immune$coefficients[sig.genes,4],treat.immune$coefficients[sig.genes,5],labels=rownames(dt.immune)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.immune$coefficients[,4],treat.immune$coefficients[,6],xlab="logFC: Fetal Male vs Female", ylab="logFC: DCM Male vs Female",main="immune cells: Sex differences - Fetal and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.immune[,4] !=0 | dt.immune[,6] != 0
text(treat.immune$coefficients[sig.genes,4],treat.immune$coefficients[sig.genes,6],labels=rownames(dt.immune)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.immune$coefficients[,5],treat.immune$coefficients[,6],xlab="logFC: ND Male vs Female", ylab="logFC: DCM Male vs Female",main="immune cells: Sex differences - ND and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
sig.genes <- dt.immune[,5] !=0 | dt.immune[,6] != 0
text(treat.immune$coefficients[sig.genes,5],treat.immune$coefficients[sig.genes,6],labels=rownames(dt.immune)[sig.genes],col=2,cex=0.6)
abline(h=0,v=0)
```


```{r}
par(mar=c(5,5,3,2))
plot(treat.immune$coefficients[,4],treat.immune$coefficients[,5],xlab="logFC: Fetal Male vs Female", ylab="logFC: ND Male vs Female",main="immune cells: Interaction term: Fetal vs ND",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.immune[,8] !=0 
if(sum(sig.genes)!=0){
  text(treat.immune$coefficients[sig.genes,4],treat.immune$coefficients[sig.genes,5],
       labels=rownames(dt.immune)[sig.genes],col=2,cex=0.6)
}
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.immune$coefficients[,4],treat.immune$coefficients[,6],xlab="logFC: Fetal Male vs Female", ylab="logFC: DCM Male vs Female",main="immune cells: Interaction term: Fetal and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.immune[,7] !=0 
if(length(rownames(dt.immune)[sig.genes]) !=0)
text(treat.immune$coefficients[sig.genes,4],treat.immune$coefficients[sig.genes,6],labels=rownames(dt.immune)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.immune$coefficients[,5],treat.immune$coefficients[,6],xlab="logFC: ND Male vs Female", ylab="logFC: DCM Male vs Female",main="immune cells: Interaction term: ND and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.immune[,9] !=0 
text(treat.immune$coefficients[sig.genes,5],treat.immune$coefficients[sig.genes,6],labels=rownames(dt.immune)[sig.genes],col=2,cex=0.6)
```

## Neurons

```{r}
cont.neurons <- makeContrasts(NvF = 0.5*(neurons.ND.Male + neurons.ND.Female) - 0.5*(neurons.Fetal.Male + neurons.Fetal.Female),
                         DvF = 0.5*(neurons.DCM.Male + neurons.DCM.Female) - 0.5*(neurons.Fetal.Male + neurons.Fetal.Female),
                         DvN = 0.5*(neurons.DCM.Male + neurons.DCM.Female) - 0.5*(neurons.ND.Male + neurons.ND.Female),
                         SexFetal = neurons.Fetal.Male - neurons.Fetal.Female,
                         SexND = neurons.ND.Male - neurons.ND.Female,
                         SexDCM = neurons.DCM.Male - neurons.DCM.Female,
                         InteractionDF = (neurons.DCM.Male - neurons.Fetal.Male) - (neurons.DCM.Female - neurons.Fetal.Female),
                         InteractionNF = (neurons.ND.Male - neurons.Fetal.Male) - (neurons.ND.Female - neurons.Fetal.Female),
                         InteractionDN = (neurons.DCM.Male - neurons.ND.Male) - (neurons.DCM.Female - neurons.ND.Female),
                         levels=design)
fit.neurons <- contrasts.fit(fit,contrasts = cont.neurons)
fit.neurons <- eBayes(fit.neurons,robust=TRUE)

summary(decideTests(fit.neurons))

treat.neurons <- treat(fit.neurons,lfc=0.5)

dt.neurons<-decideTests(treat.neurons)

summary(dt.neurons)
```

```{r}
par(mfrow=c(3,3))
for(i in 1:9){
  plotMD(treat.neurons,coef=i,status=dt.neurons[,i],hl.col=c("blue","red"))
  abline(h=0,col="grey")
}
```


```{r}
par(mfrow=c(1,1))
par(mar=c(7,4,2,2))
barplot(summary(dt.neurons)[-2,],beside=TRUE,legend=TRUE,col=c("blue","red"),ylab="Number of significant genes",las=2)
title("Neurons")
```

### ND vs Fetal
```{r}
options(digits=3)
topTreat(treat.neurons,coef=1,n=20,p.value=0.05)[,-c(1:3)]
```

### DCM vs Fetal
```{r}
topTreat(treat.neurons,coef=2,n=20,p.value=0.05)[,-c(1:3)]
```

### DCM vs ND
```{r}
topTreat(treat.neurons,coef=3)[,-c(1:3)]
```

### Male vs Female in Fetal samples
```{r}
topTreat(treat.neurons,coef=4)[,-c(1:3)]
```

### Male vs Female in ND samples
```{r}
topTreat(treat.neurons,coef=5)[,-c(1:3)]
```

### Male vs Female in DCM samples
```{r}
topTreat(treat.neurons,coef=6)[,-c(1:3)]
```

### Interaction of sex differences between DCM and Fetal
```{r}
topTreat(treat.neurons,coef=7)[,-c(1:3)]
```

### Interaction of sex differences between ND and Fetal
```{r}
topTreat(treat.neurons,coef=8)[,-c(1:3)]
```

### Interaction of sex differences between DCM and ND
```{r}
topTreat(treat.neurons,coef=9)[,-c(1:3)]
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.neurons$coefficients[,1],treat.neurons$coefficients[,2],xlab="logFC: ND vs Fetal", ylab="logFC: DCM vs Fetal",main="neurons",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0,col=colours()[c(175)])
sig.genes <-  dt.neurons[,1] !=0 | dt.neurons[,2] != 0
text(treat.neurons$coefficients[sig.genes,1],treat.neurons$coefficients[sig.genes,2],labels=rownames(dt.neurons)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.neurons$coefficients[,4],treat.neurons$coefficients[,5],xlab="logFC: Fetal Male vs Female", ylab="logFC: ND Male vs Female",main="neurons: Sex differences - Fetal and ND",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.neurons[,4] !=0 | dt.neurons[,5] != 0
text(treat.neurons$coefficients[sig.genes,4],treat.neurons$coefficients[sig.genes,5],labels=rownames(dt.neurons)[sig.genes],col=2,cex=0.6)
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.neurons$coefficients[,4],treat.neurons$coefficients[,6],xlab="logFC: Fetal Male vs Female", ylab="logFC: DCM Male vs Female",main="neurons: Sex differences - Fetal and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.neurons[,4] !=0 | dt.neurons[,6] != 0
if(sum(sig.genes)!=0){
  text(treat.neurons$coefficients[sig.genes,4],treat.neurons$coefficients[sig.genes,6],
       labels=rownames(dt.neurons)[sig.genes],col=2,cex=0.6)
}
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.neurons$coefficients[,5],treat.neurons$coefficients[,6],xlab="logFC: ND Male vs Female", ylab="logFC: DCM Male vs Female",main="neurons: Sex differences - ND and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
sig.genes <- dt.neurons[,5] !=0 | dt.neurons[,6] != 0
text(treat.neurons$coefficients[sig.genes,5],treat.neurons$coefficients[sig.genes,6],labels=rownames(dt.neurons)[sig.genes],col=2,cex=0.6)
abline(h=0,v=0)
```


```{r}
par(mar=c(5,5,3,2))
plot(treat.neurons$coefficients[,4],treat.neurons$coefficients[,5],xlab="logFC: Fetal Male vs Female", ylab="logFC: ND Male vs Female",main="neurons: Interaction term: Fetal vs ND",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.neurons[,8] !=0 
if(sum(sig.genes)!=0){
  text(treat.neurons$coefficients[sig.genes,4],treat.neurons$coefficients[sig.genes,5],
       labels=rownames(dt.neurons)[sig.genes],col=2,cex=0.6)
}
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.neurons$coefficients[,4],treat.neurons$coefficients[,6],xlab="logFC: Fetal Male vs Female", ylab="logFC: DCM Male vs Female",main="neurons: Interaction term: Fetal and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.neurons[,7] !=0 
if(sum(sig.genes)!=0){
  text(treat.neurons$coefficients[sig.genes,4],treat.neurons$coefficients[sig.genes,6],
       labels=rownames(dt.neurons)[sig.genes],col=2,cex=0.6)
}
```

```{r}
par(mar=c(5,5,3,2))
plot(treat.neurons$coefficients[,5],treat.neurons$coefficients[,6],xlab="logFC: ND Male vs Female", ylab="logFC: DCM Male vs Female",main="neurons: Interaction term: ND and DCM",col=rgb(0,0,1,alpha=0.25),pch=16,cex=0.8)
abline(h=0,v=0)
sig.genes <- dt.neurons[,9] !=0 
if(length(rownames(dt.neurons)[sig.genes]) !=0)
text(treat.neurons$coefficients[sig.genes,5],treat.neurons$coefficients[sig.genes,6],labels=rownames(dt.neurons)[sig.genes],col=2,cex=0.6)
```


# Write out results

```{rm}
fdr.cardio <- apply(treat.cardio$p.value, 2, function(x) p.adjust(x, method="BH"))
output.cardio <- data.frame(treat.cardio$genes,LogFC=treat.cardio$coefficients,AveExp=treat.cardio$Amean,tstat=treat.cardio$t, pvalue=treat.cardio$p.value, fdr=fdr.cardio)
write.csv(output.cardio,file="./output/DEAnalysis/BroadCellTypes/de-cardio-all.csv")

fdr.endo <- apply(treat.endo$p.value, 2, function(x) p.adjust(x, method="BH"))
output.endo <- data.frame(treat.endo$genes,LogFC=treat.endo$coefficients,AveExp=treat.endo$Amean,tstat=treat.endo$t, pvalue=treat.endo$p.value, fdr=fdr.endo)
write.csv(output.endo,file="./output/DEAnalysis/BroadCellTypes/de-endo-all.csv")

fdr.peri <- apply(treat.peri$p.value, 2, function(x) p.adjust(x, method="BH"))
output.peri <- data.frame(treat.peri$genes,LogFC=treat.peri$coefficients,AveExp=treat.peri$Amean,tstat=treat.peri$t, pvalue=treat.peri$p.value, fdr=fdr.peri)
write.csv(output.peri,file="./output/DEAnalysis/BroadCellTypes/de-pericyte-all.csv")

fdr.fibro <- apply(treat.fibro$p.value, 2, function(x) p.adjust(x, method="BH"))
output.fibro <- data.frame(treat.fibro$genes,LogFC=treat.fibro$coefficients,AveExp=treat.fibro$Amean,tstat=treat.fibro$t, pvalue=treat.fibro$p.value, fdr=fdr.fibro)
write.csv(output.fibro,file="./output/DEAnalysis/BroadCellTypes/de-fibro-all.csv")

fdr.immune <- apply(treat.immune$p.value, 2, function(x) p.adjust(x, method="BH"))
output.immune <- data.frame(treat.immune$genes,LogFC=treat.immune$coefficients,AveExp=treat.immune$Amean,tstat=treat.immune$t, pvalue=treat.immune$p.value, fdr=fdr.immune)
write.csv(output.immune,file="./output/DEAnalysis/BroadCellTypes/de-immune-all.csv")

fdr.neurons <- apply(treat.neurons$p.value, 2, function(x) p.adjust(x, method="BH"))
output.neurons <- data.frame(treat.neurons$genes,LogFC=treat.neurons$coefficients,AveExp=treat.neurons$Amean,tstat=treat.neurons$t, pvalue=treat.neurons$p.value, fdr=fdr.neurons)
write.csv(output.neurons,file="./output/DEAnalysis/BroadCellTypes/de-neurons-all.csv")

fdr.smc <- apply(treat.smc$p.value, 2, function(x) p.adjust(x, method="BH"))
output.smc <- data.frame(treat.smc$genes,LogFC=treat.smc$coefficients,AveExp=treat.smc$Amean,tstat=treat.smc$t, pvalue=treat.smc$p.value, fdr=fdr.smc)
write.csv(output.smc,file="./output/DEAnalysis/BroadCellTypes/de-smc-all.csv")
```

```{rm}
# Standard limma analysis with no fold change cut-off
# Cardiomyocytes
for(i in 1:ncol(fit.cardio)){
  write.csv(topTable(fit.cardio,coef=i,n="Inf"),
            file=paste("./output/DEAnalysis/BroadCellTypes/STD-ind/de-cardio-",colnames(fit.cardio)[i],".csv",sep=""),
            row.names = FALSE)
}

# Endothelium
for(i in 1:ncol(fit.endo)){
  write.csv(topTable(fit.endo,coef=i,n="Inf"),
            file=paste("./output/DEAnalysis/BroadCellTypes/STD-ind/de-endo-",colnames(fit.endo)[i],".csv",sep=""),
            row.names = FALSE)
}

# Pericyte
for(i in 1:ncol(fit.peri)){
  write.csv(topTable(fit.peri,coef=i,n="Inf"),
            file=paste("./output/DEAnalysis/BroadCellTypes/STD-ind/de-pericyte-",colnames(fit.peri)[i],".csv",sep=""),
            row.names = FALSE)
}

# Fibroblasts
for(i in 1:ncol(fit.fibro)){
  write.csv(topTable(fit.fibro,coef=i,n="Inf"),
            file=paste("./output/DEAnalysis/BroadCellTypes/STD-ind/de-fibro-",colnames(fit.fibro)[i],".csv",sep=""),
            row.names = FALSE)
}

# Immune
for(i in 1:ncol(fit.immune)){
  write.csv(topTable(fit.immune,coef=i,n="Inf"),
            file=paste("./output/DEAnalysis/BroadCellTypes/STD-ind/de-immune-",colnames(fit.immune)[i],".csv",sep=""),
            row.names = FALSE)
}

# Neurons
for(i in 1:ncol(fit.neurons)){
  write.csv(topTable(fit.neurons,coef=i,n="Inf"),
            file=paste("./output/DEAnalysis/BroadCellTypes/STD-ind/de-neurons-",colnames(fit.neurons)[i],".csv",sep=""),
            row.names = FALSE)
}

# smooth muscle cells
for(i in 1:ncol(fit.smc)){
  write.csv(topTable(fit.smc,coef=i,n="Inf"),
            file=paste("./output/DEAnalysis/BroadCellTypes/STD-ind/de-smc-",colnames(fit.smc)[i],".csv",sep=""),
            row.names = FALSE)
}


# TREAT analysis with log-fold change cut-off of 0.5
# Cardiomyocytes
for(i in 1:ncol(treat.cardio)){
  write.csv(topTreat(treat.cardio,coef=i,n="Inf"),
            file=paste("./output/DEAnalysis/BroadCellTypes/TREAT-indiv/de-cardio-",colnames(treat.cardio)[i],".csv",sep=""),
            row.names = FALSE)
}

# Endothelium
for(i in 1:ncol(treat.endo)){
  write.csv(topTreat(treat.endo,coef=i,n="Inf"),
            file=paste("./output/DEAnalysis/BroadCellTypes/TREAT-indiv/de-endo-",colnames(treat.endo)[i],".csv",sep=""),
            row.names = FALSE)
}

# Pericyte
for(i in 1:ncol(treat.peri)){
  write.csv(topTreat(treat.peri,coef=i,n="Inf"),
            file=paste("./output/DEAnalysis/BroadCellTypes/TREAT-indiv/de-pericyte-",colnames(treat.peri)[i],".csv",sep=""),
            row.names = FALSE)
}

# Fibroblasts
for(i in 1:ncol(treat.fibro)){
  write.csv(topTreat(treat.fibro,coef=i,n="Inf"),
            file=paste("./output/DEAnalysis/BroadCellTypes/TREAT-indiv/de-fibro-",colnames(treat.fibro)[i],".csv",sep=""),
            row.names = FALSE)
}

# Immune
for(i in 1:ncol(treat.immune)){
  write.csv(topTreat(treat.immune,coef=i,n="Inf"),
            file=paste("./output/DEAnalysis/BroadCellTypes/TREAT-indiv/de-immune-",colnames(treat.immune)[i],".csv",sep=""),
            row.names = FALSE)
}

# Neurons
for(i in 1:ncol(treat.neurons)){
  write.csv(topTreat(treat.neurons,coef=i,n="Inf"),
            file=paste("./output/DEAnalysis/BroadCellTypes/TREAT-indiv/de-neurons-",colnames(treat.neurons)[i],".csv",sep=""),
            row.names = FALSE)
}

# smooth muscle cells
for(i in 1:ncol(treat.smc)){
  write.csv(topTreat(treat.smc,coef=i,n="Inf"),
            file=paste("./output/DEAnalysis/BroadCellTypes/TREAT-indiv/de-smc-",colnames(treat.smc)[i],".csv",sep=""),
            row.names = FALSE)
}
```
